# Conflicts:
#	src/routes/bookingRoute.ts

+++ src/models/Booking.ts
@@ -2,7 +2,8 @@ import {PrismaClient, BookingStatus, PaymentStatus} from '@prisma/client';
 import SlotManager from './SlotManager';
 import Invoice from './Invoice';
 import INotifySubject from "../interfaces/INotifySubject";
-import INotifyObserver from "../interfaces/INotifyObserver"; // Ensure the path is correct
+import INotifyObserver from "../interfaces/INotifyObserver";
+import ParkingSlot from "./ParkingSlot"; // Ensure the path is correct
 
 const prisma = new PrismaClient();
 
@@ -20,7 +21,7 @@ class Booking implements INotifySubject {
     private slotManager: SlotManager = SlotManager.getInstance()
 
 
-    constructor(customerId: number, slotId: number, startTime: Date, endTime: Date, totalPrice: number, status: BookingStatus, id? : number) {
+    constructor(customerId: number, slotId: number, startTime: Date, endTime: Date, totalPrice: number, status: BookingStatus, id?: number) {
         this.customerId = customerId;
         this.slotId = slotId;
         this.startTime = startTime;
@@ -58,28 +59,76 @@ class Booking implements INotifySubject {
             return null
         }
     }
-    async upsertBooking(booking : Booking): Promise<boolean> {
+
+    static async getSlotById(id: number): Promise<Booking | Error> {
         try {
-            const result = await prisma.booking.upsert({
+            const booking = await prisma.booking.findUnique({
+                where: {id: id},
+            });
+            if (!booking) {
+                return Error("Booking not found");
+            } else {
+                return new Booking(booking.customerId, booking.slotId, booking.startTime, booking.endTime, booking.totalPrice, booking.status, booking.id);
+            }
+
+
+        } catch (error) {
+            console.error("Error fetching parking slot:", error);
+            return Error("Slot not found");
+        }
+    }
+
+    static async getSlotByUId(userId: number): Promise<Booking[] | Error> {
+        try {
+            const books = await prisma.booking.findMany({
+                where: {customerId: userId},
+            });
+            if (!books) {
+                return Error("Booking not found");
+            } else {
+                return books.map(booking => new Booking(booking.customerId, booking.slotId, booking.startTime, booking.endTime, booking.totalPrice, booking.status));
+            }
+        } catch (error) {
+            console.error("Error fetching booking:", error);
+            return Error("Booking not found");
+        }
+    }
+
+    static async insertBooking(booking: Booking): Promise<boolean> {
+        try {
+            const book = await prisma.booking.create({
+                data: {
+                    customerId: booking.customerId,
+                    slotId: booking.slotId,
+                    startTime: booking.startTime,
+                    endTime: booking.endTime,
+                    totalPrice: booking.totalPrice,
+                    status: booking.status,
+                },
+            });
+            const invoice = new Invoice(book.id, booking.totalPrice, new Date(), PaymentStatus.Pending);
+            await invoice.save();
+            return true;
+        } catch (error) {
+            console.error('Error upsert booking:', error);
+            return false;
+        }
+    }
+
+    static async updateBooking(booking: Booking): Promise<boolean> {
+        try {
+            const result = await prisma.booking.update({
                 where: {
                     id: booking.id || 0
                 },
-                update: {
+                data: {
                     customerId: booking.customerId,
                     slotId: booking.slotId,
                     startTime: booking.startTime,
                     endTime: booking.endTime,
                     totalPrice: booking.totalPrice,
                     status: booking.status,
-                },
-                create: {
-                    customerId: booking.customerId,
-                    slotId: booking.slotId,
-                    startTime: booking.startTime,
-                    endTime: booking.endTime,
-                    totalPrice: booking.totalPrice,
-                    status: booking.status,
-                },
+                }
             });
             return true;
         } catch (error) {
@@ -87,20 +136,26 @@ class Booking implements INotifySubject {
             return false;
         }
     }
-    async deleteBooking(booking: Booking) : Promise<boolean>
-    {
+
+    static async deleteBooking(booking: Booking): Promise<boolean> {
         try {
-            const result = await prisma.booking.delete({
-                where: {
-                    id: booking.id || 0,
-                },
-            });
+            if (booking.id) {
+                const delInvoice = await Invoice.deleteInvoice(booking.id);
+                if (delInvoice) {
+                    const result = await prisma.booking.delete({
+                        where: {
+                            id: booking.id || 0,
+                        },
+                    });
+                }
+            }
             return true;
         } catch (error) {
             console.error('Error delete parking slot:', error);
             return false;
         }
     }
+
     attach(observer: INotifyObserver): void {
         this.observers.push(observer);
     }
